+++ Laboratory/Rust/benchmarking/src/bebops/jazz.rs
@@ -16,9 +16,9 @@
 
 #![allow(warnings)]
 
-use bebop::FixedSized as _;
-use core::convert::TryInto as _;
-use std::io::Write as _;
+use ::std::io::Write as _;
+use ::core::convert::TryInto as _;
+use ::bebop::FixedSized as _;
 
 #[repr(u32)]
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
@@ -40,9 +40,7 @@ impl ::core::convert::TryFrom<u32> for Instrument {
             2 => Ok(Instrument::Clarinet),
             3 => Ok(Instrument::Piano),
             4 => Ok(Instrument::Cello),
-            d => Err(::bebop::DeserializeError::InvalidEnumDiscriminator(
-                d.into(),
-            )),
+            d => Err(::bebop::DeserializeError::InvalidEnumDiscriminator(d.into())),
         }
     }
 }
@@ -59,14 +57,12 @@ impl ::core::convert::From<Instrument> for u32 {
     }
 }
 
-impl<'raw> ::bebop::SubRecord<'raw> for Instrument {
+impl ::bebop::SubRecord<'_> for Instrument {
     const MIN_SERIALIZED_SIZE: usize = ::std::mem::size_of::<u32>();
     const EXACT_SERIALIZED_SIZE: Option<usize> = Some(::std::mem::size_of::<u32>());
 
     #[inline]
-    fn serialized_size(&self) -> usize {
-        ::std::mem::size_of::<u32>()
-    }
+    fn serialized_size(&self) -> usize { ::std::mem::size_of::<u32>() }
 
     #[inline]
     fn _serialize_chained<W: ::std::io::Write>(&self, dest: &mut W) -> ::bebop::SeResult<usize> {
@@ -74,7 +70,7 @@ impl<'raw> ::bebop::SubRecord<'raw> for Instrument {
     }
 
     #[inline]
-    fn _deserialize_chained(raw: &'raw [u8]) -> ::bebop::DeResult<(usize, Self)> {
+    fn _deserialize_chained(raw: &[u8]) -> ::bebop::DeResult<(usize, Self)> {
         let (n, v) = u32::_deserialize_chained(raw)?;
         Ok((n, v.try_into()?))
     }
@@ -84,6 +80,7 @@ impl ::bebop::FixedSized for Instrument {
     const SERIALIZED_SIZE: usize = ::std::mem::size_of::<u32>();
 }
 
+
 #[derive(Clone, Debug, PartialEq)]
 pub struct Performer<'raw> {
     pub name: &'raw str,
@@ -92,15 +89,20 @@ pub struct Performer<'raw> {
 
 impl<'raw> ::bebop::SubRecord<'raw> for Performer<'raw> {
     const MIN_SERIALIZED_SIZE: usize =
-        <&'raw str>::MIN_SERIALIZED_SIZE + <Instrument>::MIN_SERIALIZED_SIZE;
+        <&'raw str>::MIN_SERIALIZED_SIZE +
+        <Instrument>::MIN_SERIALIZED_SIZE;
 
     #[inline]
     fn serialized_size(&self) -> usize {
-        self.name.serialized_size() + self.plays.serialized_size()
+        self.name.serialized_size() +
+        self.plays.serialized_size()
     }
 
     fn _serialize_chained<W: ::std::io::Write>(&self, dest: &mut W) -> ::bebop::SeResult<usize> {
-        Ok(self.name._serialize_chained(dest)? + self.plays._serialize_chained(dest)?)
+        Ok(
+            self.name._serialize_chained(dest)? +
+            self.plays._serialize_chained(dest)?
+        )
     }
 
     fn _deserialize_chained(raw: &'raw [u8]) -> ::bebop::DeResult<(usize, Self)> {
@@ -110,18 +112,15 @@ impl<'raw> ::bebop::SubRecord<'raw> for Performer<'raw> {
             return Err(::bebop::DeserializeError::MoreDataExpected(missing));
         }
 
-        let (read, v0) = <&'raw str>::_deserialize_chained(&raw[i..])?;
+        let (read, v0) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
         i += read;
-        let (read, v1) = <Instrument>::_deserialize_chained(&raw[i..])?;
+        let (read, v1) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
         i += read;
 
-        Ok((
-            i,
-            Self {
-                name: v0,
-                plays: v1,
-            },
-        ))
+        Ok((i, Self {
+            name: v0,
+            plays: v1,
+        }))
     }
 }
 
@@ -142,23 +141,10 @@ impl<'raw> ::bebop::SubRecord<'raw> for Song<'raw> {
 
     #[inline]
     fn serialized_size(&self) -> usize {
-        ::bebop::LEN_SIZE
-            + 1
-            + self
-                .title
-                .as_ref()
-                .map(|v| v.serialized_size() + 1)
-                .unwrap_or(0)
-            + self
-                .year
-                .as_ref()
-                .map(|v| v.serialized_size() + 1)
-                .unwrap_or(0)
-            + self
-                .performers
-                .as_ref()
-                .map(|v| v.serialized_size() + 1)
-                .unwrap_or(0)
+        ::bebop::LEN_SIZE + 1 +
+        self.title.as_ref().map(|v| v.serialized_size() + 1).unwrap_or(0) +
+        self.year.as_ref().map(|v| v.serialized_size() + 1).unwrap_or(0) +
+        self.performers.as_ref().map(|v| v.serialized_size() + 1).unwrap_or(0)
     }
 
     fn _serialize_chained<W: ::std::io::Write>(&self, dest: &mut W) -> ::bebop::SeResult<usize> {
@@ -222,7 +208,7 @@ impl<'raw> ::bebop::SubRecord<'raw> for Song<'raw> {
                     if _title.is_some() {
                         return Err(::bebop::DeserializeError::DuplicateMessageField);
                     }
-                    let (read, value) = <&'raw str>::_deserialize_chained(&raw[i..])?;
+                    let (read, value) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
                     i += read;
                     _title = Some(value)
                 }
@@ -231,7 +217,7 @@ impl<'raw> ::bebop::SubRecord<'raw> for Song<'raw> {
                     if _year.is_some() {
                         return Err(::bebop::DeserializeError::DuplicateMessageField);
                     }
-                    let (read, value) = <u16>::_deserialize_chained(&raw[i..])?;
+                    let (read, value) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
                     i += read;
                     _year = Some(value)
                 }
@@ -240,8 +226,7 @@ impl<'raw> ::bebop::SubRecord<'raw> for Song<'raw> {
                     if _performers.is_some() {
                         return Err(::bebop::DeserializeError::DuplicateMessageField);
                     }
-                    let (read, value) =
-                        <::std::vec::Vec<Performer<'raw>>>::_deserialize_chained(&raw[i..])?;
+                    let (read, value) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
                     i += read;
                     _performers = Some(value)
                 }
@@ -254,17 +239,14 @@ impl<'raw> ::bebop::SubRecord<'raw> for Song<'raw> {
 
         if i != len {
             debug_assert!(i > len);
-            return Err(::bebop::DeserializeError::CorruptFrame);
+            return Err(::bebop::DeserializeError::CorruptFrame)
         }
 
-        Ok((
-            i,
-            Self {
-                title: _title,
-                year: _year,
-                performers: _performers,
-            },
-        ))
+        Ok((i, Self {
+            title: _title,
+            year: _year,
+            performers: _performers,
+        }))
     }
 }
 
@@ -276,7 +258,9 @@ pub enum Album<'raw> {
     Unknown,
 
     /// Discriminator 1
-    StudioAlbum { tracks: ::std::vec::Vec<Song<'raw>> },
+    StudioAlbum {
+        tracks: ::std::vec::Vec<Song<'raw>>,
+    },
 
     /// Discriminator 2
     LiveAlbum {
@@ -293,34 +277,29 @@ impl<'raw> ::bebop::SubRecord<'raw> for Album<'raw> {
     const MIN_SERIALIZED_SIZE: usize = ::bebop::LEN_SIZE + 1;
 
     fn serialized_size(&self) -> usize {
-        ::bebop::LEN_SIZE
-            + 1
-            + match self {
-                Album::Unknown => 0,
-                Album::StudioAlbum {
-                    tracks: ref _tracks,
-                } => _tracks.serialized_size(),
-                Album::LiveAlbum {
-                    tracks: ref _tracks,
-                    venue_name: ref _venue_name,
-                    concert_date: ref _concert_date,
-                } => {
-                    ::bebop::LEN_SIZE
-                        + 1
-                        + _tracks
-                            .as_ref()
-                            .map(|v| v.serialized_size() + 1)
-                            .unwrap_or(0)
-                        + _venue_name
-                            .as_ref()
-                            .map(|v| v.serialized_size() + 1)
-                            .unwrap_or(0)
-                        + _concert_date
-                            .as_ref()
-                            .map(|v| v.serialized_size() + 1)
-                            .unwrap_or(0)
-                }
+        ::bebop::LEN_SIZE + 1 +
+        match self {
+            Album::Unknown => {
+                0
             }
+            Self::StudioAlbum {
+                tracks: ref _tracks,
+            }
+            => {
+                _tracks.serialized_size()
+            }
+            Self::LiveAlbum {
+                tracks: ref _tracks,
+                venue_name: ref _venue_name,
+                concert_date: ref _concert_date,
+            }
+            => {
+                ::bebop::LEN_SIZE + 1 +
+                _tracks.as_ref().map(|v| v.serialized_size() + 1).unwrap_or(0) +
+                _venue_name.as_ref().map(|v| v.serialized_size() + 1).unwrap_or(0) +
+                _concert_date.as_ref().map(|v| v.serialized_size() + 1).unwrap_or(0)
+            }
+        }
     }
 
     fn _serialize_chained<W: ::std::io::Write>(&self, dest: &mut W) -> ::bebop::SeResult<usize> {
@@ -330,17 +309,19 @@ impl<'raw> ::bebop::SubRecord<'raw> for Album<'raw> {
             Album::Unknown => {
                 return Err(::bebop::SerializeError::CannotSerializeUnknownUnion);
             }
-            Album::StudioAlbum {
+            Self::StudioAlbum {
                 tracks: ref _tracks,
-            } => {
+            }
+            => {
                 1u8._serialize_chained(dest)?;
                 _tracks._serialize_chained(dest)?;
             }
-            Album::LiveAlbum {
+            Self::LiveAlbum {
                 tracks: ref _tracks,
                 venue_name: ref _venue_name,
                 concert_date: ref _concert_date,
-            } => {
+            }
+            => {
                 2u8._serialize_chained(dest)?;
                 ::bebop::write_len(dest, size - ::bebop::LEN_SIZE * 2 - 1)?;
                 if let Some(ref v) = _tracks {
@@ -366,10 +347,12 @@ impl<'raw> ::bebop::SubRecord<'raw> for Album<'raw> {
         let mut i = ::bebop::LEN_SIZE + 1;
         let de = match raw[::bebop::LEN_SIZE] {
             1 => {
-                let (read, v0) = <::std::vec::Vec<Song<'raw>>>::_deserialize_chained(&raw[i..])?;
+                let (read, v0) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
                 i += read;
 
-                Album::StudioAlbum { tracks: v0 }
+                Album::StudioAlbum {
+                    tracks: v0,
+                }
             }
             2 => {
                 let len = ::bebop::read_len(&raw[i..])? + i + ::bebop::LEN_SIZE;
@@ -412,8 +395,7 @@ impl<'raw> ::bebop::SubRecord<'raw> for Album<'raw> {
                             if _tracks.is_some() {
                                 return Err(::bebop::DeserializeError::DuplicateMessageField);
                             }
-                            let (read, value) =
-                                <::std::vec::Vec<Song<'raw>>>::_deserialize_chained(&raw[i..])?;
+                            let (read, value) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
                             i += read;
                             _tracks = Some(value)
                         }
@@ -422,7 +404,7 @@ impl<'raw> ::bebop::SubRecord<'raw> for Album<'raw> {
                             if _venue_name.is_some() {
                                 return Err(::bebop::DeserializeError::DuplicateMessageField);
                             }
-                            let (read, value) = <&'raw str>::_deserialize_chained(&raw[i..])?;
+                            let (read, value) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
                             i += read;
                             _venue_name = Some(value)
                         }
@@ -431,7 +413,7 @@ impl<'raw> ::bebop::SubRecord<'raw> for Album<'raw> {
                             if _concert_date.is_some() {
                                 return Err(::bebop::DeserializeError::DuplicateMessageField);
                             }
-                            let (read, value) = <::bebop::Date>::_deserialize_chained(&raw[i..])?;
+                            let (read, value) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
                             i += read;
                             _concert_date = Some(value)
                         }
@@ -444,7 +426,7 @@ impl<'raw> ::bebop::SubRecord<'raw> for Album<'raw> {
 
                 if i != len {
                     debug_assert!(i > len);
-                    return Err(::bebop::DeserializeError::CorruptFrame);
+                    return Err(::bebop::DeserializeError::CorruptFrame)
                 }
 
                 Album::LiveAlbum {
@@ -461,10 +443,12 @@ impl<'raw> ::bebop::SubRecord<'raw> for Album<'raw> {
         if !cfg!(feature = "unchecked") && i != len {
             debug_assert!(i > len);
             Err(::bebop::DeserializeError::CorruptFrame)
-        } else {
+        }
+        else {
             Ok((i, de))
         }
     }
+
 }
 
 impl<'raw> ::bebop::Record<'raw> for Album<'raw> {}
@@ -484,7 +468,9 @@ impl<'raw> ::bebop::SubRecord<'raw> for Library<'raw> {
     }
 
     fn _serialize_chained<W: ::std::io::Write>(&self, dest: &mut W) -> ::bebop::SeResult<usize> {
-        Ok(self.albums._serialize_chained(dest)?)
+        Ok(
+            self.albums._serialize_chained(dest)?
+        )
     }
 
     fn _deserialize_chained(raw: &'raw [u8]) -> ::bebop::DeResult<(usize, Self)> {
@@ -494,12 +480,485 @@ impl<'raw> ::bebop::SubRecord<'raw> for Library<'raw> {
             return Err(::bebop::DeserializeError::MoreDataExpected(missing));
         }
 
-        let (read, v0) =
-            <::std::collections::HashMap<&'raw str, Album<'raw>>>::_deserialize_chained(&raw[i..])?;
+        let (read, v0) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
         i += read;
 
-        Ok((i, Self { albums: v0 }))
+        Ok((i, Self {
+            albums: v0,
+        }))
     }
 }
 
 impl<'raw> ::bebop::Record<'raw> for Library<'raw> {}
+
+#[cfg(feature = "bebop-owned-all")]
+pub mod owned {
+    #![allow(warnings)]
+
+    use ::std::io::Write as _;
+    use ::core::convert::TryInto as _;
+    use ::bebop::FixedSized as _;
+
+    pub use super::Instrument;#[derive(Clone, Debug, PartialEq)]
+    pub struct Performer {
+        pub name: String,
+        pub plays: Instrument,
+    }
+
+    impl<'raw> ::core::convert::From<super::Performer<'raw>> for Performer {
+        fn from(value: super::Performer) -> Self {
+            Self {
+                name: value.name.into(),
+                plays: value.plays,
+            }
+        }
+    }
+
+    impl<'raw> ::bebop::SubRecord<'raw> for Performer {
+        const MIN_SERIALIZED_SIZE: usize =
+            <String>::MIN_SERIALIZED_SIZE +
+            <Instrument>::MIN_SERIALIZED_SIZE;
+
+        #[inline]
+        fn serialized_size(&self) -> usize {
+            self.name.serialized_size() +
+            self.plays.serialized_size()
+        }
+
+        fn _serialize_chained<W: ::std::io::Write>(&self, dest: &mut W) -> ::bebop::SeResult<usize> {
+            Ok(
+                self.name._serialize_chained(dest)? +
+                self.plays._serialize_chained(dest)?
+            )
+        }
+
+        fn _deserialize_chained(raw: &'raw [u8]) -> ::bebop::DeResult<(usize, Self)> {
+            let mut i = 0;
+            if raw.len() - i < Self::MIN_SERIALIZED_SIZE {
+                let missing = Self::MIN_SERIALIZED_SIZE - (raw.len() - i);
+                return Err(::bebop::DeserializeError::MoreDataExpected(missing));
+            }
+
+            let (read, v0) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
+            i += read;
+            let (read, v1) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
+            i += read;
+
+            Ok((i, Self {
+                name: v0,
+                plays: v1,
+            }))
+        }
+    }
+
+    impl<'raw> ::bebop::Record<'raw> for Performer {}
+    #[derive(Clone, Debug, PartialEq, Default)]
+    pub struct {
+        /// Field 1
+        pub title: ::core::option::Option<String>,
+        /// Field 2
+        pub year: ::core::option::Option<u16>,
+        /// Field 3
+        pub performers: ::core::option::Option<::std::vec::Vec<Performer<'raw>>>,
+    }
+
+    impl<'raw> ::core::convert::From<super::Song<'raw>> for Song {
+        fn from(value: super::Song) -> Self {
+            Self {
+                title: value.title.map(|value| value.into()),
+                year: value.year,
+                performers: value.performers.map(|value| value.iter().map(|value| value.into()).collect()),
+            }
+        }
+    }
+
+    impl<'raw> ::bebop::SubRecord<'raw> for {
+        const MIN_SERIALIZED_SIZE: usize = ::bebop::LEN_SIZE + 1;
+
+        #[inline]
+        fn serialized_size(&self) -> usize {
+            ::bebop::LEN_SIZE + 1 +
+            self.title.as_ref().map(|v| v.serialized_size() + 1).unwrap_or(0) +
+            self.year.as_ref().map(|v| v.serialized_size() + 1).unwrap_or(0) +
+            self.performers.as_ref().map(|v| v.serialized_size() + 1).unwrap_or(0)
+        }
+
+        fn _serialize_chained<W: ::std::io::Write>(&self, dest: &mut W) -> ::bebop::SeResult<usize> {
+            let size = self.serialized_size();
+            ::bebop::write_len(dest, size - ::bebop::LEN_SIZE)?;
+            if let Some(ref v) = self.title {
+                1u8._serialize_chained(dest)?;
+                v._serialize_chained(dest)?;
+            }
+            if let Some(ref v) = self.year {
+                2u8._serialize_chained(dest)?;
+                v._serialize_chained(dest)?;
+            }
+            if let Some(ref v) = self.performers {
+                3u8._serialize_chained(dest)?;
+                v._serialize_chained(dest)?;
+            }
+            0u8._serialize_chained(dest)?;
+            Ok(size)
+        }
+
+        fn _deserialize_chained(raw: &'raw [u8]) -> ::bebop::DeResult<(usize, Self)> {
+            let mut i = 0;
+            let len = ::bebop::read_len(&raw[i..])? + ::bebop::LEN_SIZE;
+            i += ::bebop::LEN_SIZE;
+
+            #[cfg(not(feature = "unchecked"))]
+            if len == 0 {
+                return Err(::bebop::DeserializeError::CorruptFrame);
+            }
+
+            if raw.len() < len {
+                return Err(::bebop::DeserializeError::MoreDataExpected(len - raw.len()));
+            }
+
+            let mut _title = None;
+            let mut _year = None;
+            let mut _performers = None;
+
+            #[cfg(not(feature = "unchecked"))]
+            let mut last = 0;
+
+            while i < len {
+                let di = raw[i];
+
+                #[cfg(not(feature = "unchecked"))]
+                if di != 0 {
+                    if di < last {
+                        return Err(::bebop::DeserializeError::CorruptFrame);
+                    }
+                    last = di;
+                }
+
+                i += 1;
+                match di {
+                    0 => {
+                        break;
+                    }
+                    1 => {
+                        #[cfg(not(feature = "unchecked"))]
+                        if _title.is_some() {
+                            return Err(::bebop::DeserializeError::DuplicateMessageField);
+                        }
+                        let (read, value) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
+                        i += read;
+                        _title = Some(value)
+                    }
+                    2 => {
+                        #[cfg(not(feature = "unchecked"))]
+                        if _year.is_some() {
+                            return Err(::bebop::DeserializeError::DuplicateMessageField);
+                        }
+                        let (read, value) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
+                        i += read;
+                        _year = Some(value)
+                    }
+                    3 => {
+                        #[cfg(not(feature = "unchecked"))]
+                        if _performers.is_some() {
+                            return Err(::bebop::DeserializeError::DuplicateMessageField);
+                        }
+                        let (read, value) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
+                        i += read;
+                        _performers = Some(value)
+                    }
+                    _ => {
+                        i = len;
+                        break;
+                    }
+                }
+            }
+
+            if i != len {
+                debug_assert!(i > len);
+                return Err(::bebop::DeserializeError::CorruptFrame)
+            }
+
+            Ok((i, Self {
+                title: _title,
+                year: _year,
+                performers: _performers,
+            }))
+        }
+    }
+
+    impl<'raw> ::bebop::Record<'raw> for  {}
+    #[derive(Clone, Debug, PartialEq)]
+    pub enum {
+        /// An unknown type which is likely defined in a newer version of the schema.
+        Unknown,
+
+        /// Discriminator 1
+        StudioAlbum {
+            tracks: ::std::vec::Vec<Song<'raw>>,
+        },
+
+        /// Discriminator 2
+        LiveAlbum {
+            /// Field 1
+            tracks: ::core::option::Option<::std::vec::Vec<Song<'raw>>>,
+            /// Field 2
+            venue_name: ::core::option::Option<String>,
+            /// Field 3
+            concert_date: ::core::option::Option<::bebop::Date>,
+        },
+    }
+
+    impl<'raw> ::core::convert::From<super::Album<'raw>> for Album {
+        fn from(value: super::Album) -> Self {
+            match self {
+                Album::Unknown => {
+                    Self::Unknown
+                }
+                super::Album::StudioAlbum {
+                    tracks: _tracks,
+                }
+                => {
+                    Self::StudioAlbum {
+                        tracks: value.tracks.iter().map(|value| value.into()).collect(),
+                    }
+                }
+                super::Album::LiveAlbum {
+                    tracks: _tracks,
+                    venue_name: _venue_name,
+                    concert_date: _concert_date,
+                }
+                => {
+                    Self::LiveAlbum {
+                        tracks: value.tracks.map(|value| value.iter().map(|value| value.into()).collect()),
+                        venue_name: value.venue_name.map(|value| value.into()),
+                        concert_date: value.concert_date,
+                    }
+                }
+            }
+        }
+
+    }
+    impl<'raw> ::bebop::SubRecord<'raw> for {
+        const MIN_SERIALIZED_SIZE: usize = ::bebop::LEN_SIZE + 1;
+
+        fn serialized_size(&self) -> usize {
+            ::bebop::LEN_SIZE + 1 +
+            match self {
+                Album::Unknown => {
+                    0
+                }
+                Self::StudioAlbum {
+                    tracks: ref _tracks,
+                }
+                => {
+                    _tracks.serialized_size()
+                }
+                Self::LiveAlbum {
+                    tracks: ref _tracks,
+                    venue_name: ref _venue_name,
+                    concert_date: ref _concert_date,
+                }
+                => {
+                    ::bebop::LEN_SIZE + 1 +
+                    _tracks.as_ref().map(|v| v.serialized_size() + 1).unwrap_or(0) +
+                    _venue_name.as_ref().map(|v| v.serialized_size() + 1).unwrap_or(0) +
+                    _concert_date.as_ref().map(|v| v.serialized_size() + 1).unwrap_or(0)
+                }
+            }
+        }
+
+        fn _serialize_chained<W: ::std::io::Write>(&self, dest: &mut W) -> ::bebop::SeResult<usize> {
+            let size = self.serialized_size();
+            ::bebop::write_len(dest, size - ::bebop::LEN_SIZE - 1)?;
+            match self {
+                Album::Unknown => {
+                    return Err(::bebop::SerializeError::CannotSerializeUnknownUnion);
+                }
+                Self::StudioAlbum {
+                    tracks: ref _tracks,
+                }
+                => {
+                    1u8._serialize_chained(dest)?;
+                    _tracks._serialize_chained(dest)?;
+                }
+                Self::LiveAlbum {
+                    tracks: ref _tracks,
+                    venue_name: ref _venue_name,
+                    concert_date: ref _concert_date,
+                }
+                => {
+                    2u8._serialize_chained(dest)?;
+                    ::bebop::write_len(dest, size - ::bebop::LEN_SIZE * 2 - 1)?;
+                    if let Some(ref v) = _tracks {
+                        1u8._serialize_chained(dest)?;
+                        v._serialize_chained(dest)?;
+                    }
+                    if let Some(ref v) = _venue_name {
+                        2u8._serialize_chained(dest)?;
+                        v._serialize_chained(dest)?;
+                    }
+                    if let Some(ref v) = _concert_date {
+                        3u8._serialize_chained(dest)?;
+                        v._serialize_chained(dest)?;
+                    }
+                    0u8._serialize_chained(dest)?;
+                }
+            }
+            Ok(size)
+        }
+
+        fn _deserialize_chained(raw: &'raw [u8]) -> ::bebop::DeResult<(usize, Self)> {
+            let len = ::bebop::read_len(&raw)? + ::bebop::LEN_SIZE + 1;
+            let mut i = ::bebop::LEN_SIZE + 1;
+            let de = match raw[::bebop::LEN_SIZE] {
+                1 => {
+                    let (read, v0) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
+                    i += read;
+
+                    Album::StudioAlbum {
+                        tracks: v0,
+                    }
+                }
+                2 => {
+                    let len = ::bebop::read_len(&raw[i..])? + i + ::bebop::LEN_SIZE;
+                    i += ::bebop::LEN_SIZE;
+
+                    #[cfg(not(feature = "unchecked"))]
+                    if len == 0 {
+                        return Err(::bebop::DeserializeError::CorruptFrame);
+                    }
+
+                    if raw.len() < len {
+                        return Err(::bebop::DeserializeError::MoreDataExpected(len - raw.len()));
+                    }
+
+                    let mut _tracks = None;
+                    let mut _venue_name = None;
+                    let mut _concert_date = None;
+
+                    #[cfg(not(feature = "unchecked"))]
+                    let mut last = 0;
+
+                    while i < len {
+                        let di = raw[i];
+
+                        #[cfg(not(feature = "unchecked"))]
+                        if di != 0 {
+                            if di < last {
+                                return Err(::bebop::DeserializeError::CorruptFrame);
+                            }
+                            last = di;
+                        }
+
+                        i += 1;
+                        match di {
+                            0 => {
+                                break;
+                            }
+                            1 => {
+                                #[cfg(not(feature = "unchecked"))]
+                                if _tracks.is_some() {
+                                    return Err(::bebop::DeserializeError::DuplicateMessageField);
+                                }
+                                let (read, value) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
+                                i += read;
+                                _tracks = Some(value)
+                            }
+                            2 => {
+                                #[cfg(not(feature = "unchecked"))]
+                                if _venue_name.is_some() {
+                                    return Err(::bebop::DeserializeError::DuplicateMessageField);
+                                }
+                                let (read, value) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
+                                i += read;
+                                _venue_name = Some(value)
+                            }
+                            3 => {
+                                #[cfg(not(feature = "unchecked"))]
+                                if _concert_date.is_some() {
+                                    return Err(::bebop::DeserializeError::DuplicateMessageField);
+                                }
+                                let (read, value) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
+                                i += read;
+                                _concert_date = Some(value)
+                            }
+                            _ => {
+                                i = len;
+                                break;
+                            }
+                        }
+                    }
+
+                    if i != len {
+                        debug_assert!(i > len);
+                        return Err(::bebop::DeserializeError::CorruptFrame)
+                    }
+
+                    Album::LiveAlbum {
+                        tracks: _tracks,
+                        venue_name: _venue_name,
+                        concert_date: _concert_date,
+                    }
+                }
+                _ => {
+                    i = len;
+                    Album::Unknown
+                }
+            };
+            if !cfg!(feature = "unchecked") && i != len {
+                debug_assert!(i > len);
+                Err(::bebop::DeserializeError::CorruptFrame)
+            }
+            else {
+                Ok((i, de))
+            }
+        }
+
+    }
+
+    impl<'raw> ::bebop::Record<'raw> for  {}
+    #[derive(Clone, Debug, PartialEq)]
+    pub struct Library {
+        pub albums: ::std::collections::HashMap<String, Album<'raw>>,
+    }
+
+    impl<'raw> ::core::convert::From<super::Library<'raw>> for Library {
+        fn from(value: super::Library) -> Self {
+            Self {
+                albums: value.albums.into_iter(|(key, value)| (key.into(), value.into())).collect(),
+            }
+        }
+    }
+
+    impl<'raw> ::bebop::SubRecord<'raw> for Library {
+        const MIN_SERIALIZED_SIZE: usize =
+            <::std::collections::HashMap<String, Album<'raw>>>::MIN_SERIALIZED_SIZE;
+
+        #[inline]
+        fn serialized_size(&self) -> usize {
+            self.albums.serialized_size()
+        }
+
+        fn _serialize_chained<W: ::std::io::Write>(&self, dest: &mut W) -> ::bebop::SeResult<usize> {
+            Ok(
+                self.albums._serialize_chained(dest)?
+            )
+        }
+
+        fn _deserialize_chained(raw: &'raw [u8]) -> ::bebop::DeResult<(usize, Self)> {
+            let mut i = 0;
+            if raw.len() - i < Self::MIN_SERIALIZED_SIZE {
+                let missing = Self::MIN_SERIALIZED_SIZE - (raw.len() - i);
+                return Err(::bebop::DeserializeError::MoreDataExpected(missing));
+            }
+
+            let (read, v0) = ::bebop::SubRecord::_deserialize_chained(&raw[i..])?;
+            i += read;
+
+            Ok((i, Self {
+                albums: v0,
+            }))
+        }
+    }
+
+    impl<'raw> ::bebop::Record<'raw> for Library {}}
