+++ Laboratory/Rust/manual-testing/Cargo.toml
@@ -10,5 +10,6 @@ bebop = { path = "../../../Runtime/Rust", default-features = false }
 static_assertions = "1.1.0"
 
 [features]
-default = ["bebop-owned-all"]
+default = ["bebop-owned-all", "owned-generation"]
 bebop-owned-all = []
+owned-generation = []
